+++ config/environments/development.rb
@@ -1,7 +1,7 @@
 require "active_support/core_ext/integer/time"
 
 Rails.application.configure do
-  config.hosts << "29ae-123-254-11-172.ngrok-free.app"  # Settings specified here will take precedence over those in config/application.rb.
+  config.hosts << "306c-123-254-11-172.ngrok-free.app"  # Settings specified here will take precedence over those in config/application.rb.
 
   # In the development environment your application's code is reloaded any time
   # it changes. This slows down response time but is perfect for development
